+++ octopus/src/main/kotlin/com/koupper/octopus/annotations/AnnotationsResolvers.kt
@@ -1,9 +1,9 @@
 package com.koupper.octopus.annotations
 
-import com.koupper.octopus.logging.LogSpec
+import com.koupper.logging.LogSpec
 
 object LoggerAnnotationResolver : AnnotationResolver {
-    override fun prepare(
+    override fun prepareParams(
         scriptPath: String?,
         baseParams: MutableMap<String, Any>,
         annotationParams: Map<String, Any>,
@@ -11,7 +11,7 @@ object LoggerAnnotationResolver : AnnotationResolver {
         context: String
     ) {
         val level = (annotationParams["level"] ?: "INFO").toString()
-        val dest  = (annotationParams["destination"] ?: "file").toString()
+        val dest  = (annotationParams["destination"] ?: "console").toString()
 
         baseParams["logSpec"] = LogSpec(
             level = level,
@@ -23,7 +23,7 @@ object LoggerAnnotationResolver : AnnotationResolver {
 }
 
 object JobsListenerAnnotationResolver : AnnotationResolver {
-    override fun prepare(
+    override fun prepareParams(
         scriptPath: String?,
         baseParams: MutableMap<String, Any>,
         annotationParams: Map<String, Any>,
@@ -34,11 +34,62 @@ object JobsListenerAnnotationResolver : AnnotationResolver {
     }
 }
 
+object TimerAnnotationResolver : AnnotationResolver {
+    override fun prepareParams(
+        scriptPath: String?,
+        baseParams: MutableMap<String, Any>,
+        annotationParams: Map<String, Any>,
+        sentence: String,
+        context: String
+    ) {
+        val interval     = (annotationParams["interval"] ?: "").toString()
+        val initialDelay = (annotationParams["initialDelay"] ?: "").toString()
+        val repeat       = (annotationParams["repeat"] ?: true).toString()
+
+        baseParams["timerParams"] = mapOf(
+            "interval"     to interval,
+            "initialDelay" to initialDelay,
+            "repeat"       to repeat,
+            "scriptPath"   to (scriptPath ?: "n/a"),
+            "sentence"     to sentence
+        )
+    }
+}
+
+object ScheduleAnnotationResolver : AnnotationResolver {
+    override fun prepareParams(
+        scriptPath: String?,
+        baseParams: MutableMap<String, Any>,
+        annotationParams: Map<String, Any>,
+        sentence: String,
+        context: String
+    ) {
+        val at         = (annotationParams["at"] ?: "").toString()
+        val zone       = (annotationParams["zone"] ?: "").toString()
+        val skipIfPast = (annotationParams["skipIfPast"] ?: true).toString()
+
+        baseParams["scheduleParams"] = mapOf(
+            "at"         to at,
+            "zone"       to zone,
+            "skipIfPast" to skipIfPast,
+            "scriptPath" to (scriptPath ?: "n/a"),
+            "sentence"   to sentence
+        )
+    }
+}
+
 val annotationResolvers: Map<String, AnnotationResolver> = mapOf(
-    "Export"       to object : AnnotationResolver { // Export no prepara nada
-        override fun prepare(sp: String?, bp: MutableMap<String, Any>, ap: Map<String, Any>, s: String, c: String) {}
+    "Export"       to object : AnnotationResolver {
+        override fun prepareParams(
+            sp: String?,
+            bp: MutableMap<String, Any>,
+            ap: Map<String, Any>,
+            s: String,
+            c: String
+        ) { /* no-op */ }
     },
     "Logger"       to LoggerAnnotationResolver,
-    "JobsListener" to JobsListenerAnnotationResolver
+    "JobsListener" to JobsListenerAnnotationResolver,
+    "Timer"        to TimerAnnotationResolver,
+    "Schedule"     to ScheduleAnnotationResolver
 )
-